Move project components inside the component folder
