+++ packages/client/hmi-client/src/components/project/tera-notes-sidebar.vue
