+++ packages/client/hmi-client/src/components/project/tera-resource-sidebar.vue
