+++ packages/client/hmi-client/src/components/project/tera-upload-resources-modal.vue
