+++ packages/client/hmi-client/src/page/project/tera-project.vue
@@ -77,9 +77,9 @@ import TeraModel from '@/components/model/tera-model.vue';
 import TeraSliderPanel from '@/components/widgets/tera-slider-panel.vue';
 import TeraWorkflow from '@/components/workflow/tera-workflow.vue';
 import { useProjects } from '@/composables/project';
-import TeraNotesSidebar from '@/page/project/components/tera-notes-sidebar.vue';
-import TeraProjectOverview from '@/page/project/components/tera-project-overview.vue';
-import TeraResourceSidebar from '@/page/project/components/tera-resource-sidebar.vue';
+import TeraNotesSidebar from '@/components/project/tera-notes-sidebar.vue';
+import TeraProjectOverview from '@/components/project/tera-project-overview.vue';
+import TeraResourceSidebar from '@/components/project/tera-resource-sidebar.vue';
 import { RouteName } from '@/router/routes';
 import { ProjectPages, isProjectAssetTypes } from '@/types/Project';
 import { AssetType } from '@/types/Types';
@@ -89,8 +89,8 @@ import { isEqual } from 'lodash';
 import Button from 'primevue/button';
 import { computed, onMounted, ref } from 'vue';
 import { useRoute, useRouter } from 'vue-router';
-import TeraSaveAssetModal from './components/tera-save-asset-modal.vue';
-import TeraUploadResourcesModal from './components/tera-upload-resources-modal.vue';
+import TeraSaveAssetModal from '@/components/project/tera-save-asset-modal.vue';
+import TeraUploadResourcesModal from '@/components/project/tera-upload-resources-modal.vue';
 
 const route = useRoute();
 const router = useRouter();
